Remove build/root/usr/local/bin/entrypoint
In the Prow CI environment, the container's ENTRYPOINT is never
executed (so the Pod is launched with `spec.containers.command` set),
so no reason to keep this `entrypoint` file here.

Instead, set `INSIDE_CI_IMAGE=y` in the Dockerfile and inside `run`
command, print out all the relevant context information:
- git show
- oc version
- $0 $@ of the `run script`

and activate `set -x` to know for sure how `ansible` is launched.