roles/openshift_node: make more generic (remove GPU-specific aspect)
> toolbox/scaleup_cluster.sh [new machine type]

This command scales up the cluster using the optional machine type
passed in parameter (the default value is controlled via
`machine_instance_type` in `group_var/all.yml`).

By "scaling up" here, we mean: ensure that a Machine with the given
machine type is available in cluster.

The name of the generated `MachineSet` is also updated to include the
machine type name, so that the cluster can easily be scaled-up with
different machine types.