Change scaleup to set_scale - supported scale other than just 1 node
First step for fixing #129

Rename scaleup role to ensure_machineset which makes sure there's at-least one worker machineset with a given instance type.
If there's no such machineset, it'll create one out of an existing worker machineset (like it does now), but now with 0 replicas instead of 1.

Added new role, called set_scale which receives an instance type and a scale param which:

Ensures that the sum of spec.replicas of all machinesets with the given instance type is equal to the scale param.

If it's not, scales down all of them to 0 except for the first one which will have its spec.replicas set to the scale param.

In any case, it waits for all of machinesets to have:
`status.availableReplicas == spec.replicas`
and
`status.replicas == spec.replicas`