CI: test_master_branch: use test_commit instead of test_master_branch
`test_master_branch` uses our OLM bundle + operator image
pre/automatically built on Quay.io, but the bundle image still points
to NVIDIA v1.7.1 version of the operator.

I can't find a way to successfully fix the image on the fly, so this
commit switches the `master` branch testing to a "commit" test,
pointing to the GPU Operator master branch. This will build the
operator image in the cluster + the bundle image with the correct
operator image address.