Add role and playbook to check NFD CRD status
Add role and playbook to ensure that the NFD CRD status updates
are working properly.

Currently, the roles added in this commit ensure that NFD is already
up and running. If NFD is not up and running, then there is no sense
in trying to test the NFD CRD status output.