Replaced the bulk of the toolbox directory with Python fire CLI
Maintaining and creating new toolbox commands has become rather tedious
because of the awkward syntax of bash. Especially when trying to add
multiple optional flags to a single command. This PR replaces most
of the toolbox shell scripts with simpler Python files all launched
via the `./run.py` script at the root of the repository.

Each command (sro, nfd, gpu_operator, cluster, etc) is specified
as a Python class in its own file under the `toolbox` directory. Each subcommand
is specified as a static method of that class. The parameters for the subcommands
are specified as regular Python method parameters. All those classes must be listed
inside `./run.py`. Fire then automatically discovers all the methods of
those classes and generates a CLI.

The benefit of using Fire is that we get automatic arg parsing with free `--help`
which shows the available commands, available subcommands for each command, the
parameters, which parameters are required (positional) and which are optional (flags),
it shows the docstring of the subcommand method as the description for the subcommand.

I've also added plumbing to run ansible playbooks with ansible
parameters passed as a simple Python dictionary, so the code to create
a subcommand is almost as minimal as possible.