Make playbook role as version agnostic as possible
Currently, only OCP >= 4.8 has the status checking feature, so
earlier versions of the cluster-nfd-operator do not need to be
considered when checking resources.

The role attempts to clone a git repo (which, yes, needs to be
removed by a later role, but this is a WIP PR). Then, it parses
the YAML files to determine resource names and types.

In the next part of this WIP, I will be adding code to label all
manifests by the type of resource they create, rather than only
labeling the DaemonSet manifests in the 'daemonset_init.yml' file.

Finally, some roles are commented out for now in 'main.yml'. They
will be uncommented when they are fixed.