Name unnamed debug tasks; fix empty string comparison to use return code
Some debug tasks were unnamed, so they are now named. Additionally,
an empty string comparison was made when checking the controller
manager pod for NFD. Rather than checking stdout, stderr, etc., it
now checks the return code, where a return code of 0 indicates that
the output of said task was able to find an error message.