CI: refactor entrypoints location
This commit refactors the location of the CI entrypoints.

One goal is to be able to trigger CI tests locally, without having
issues with `PATH` references (files used to be stored in
`/usr/local/bin`.

The other goal is to be able to trigger custom tests in the CI, and
store permanently the test routes. Such tests will be specified with
the following anchor in the last commit of the PR:

test-path: testing/custom/test.sh

Note that `test-path` must point to a script in `testing/` (to avoid
running arbitrary commands not shown in the PR file diff).

The execution of *all* the test paths will be triggered with `\test
custom` in the PR message.