gpu_operator_bundle_from_commit: properly generate the CSV spec.relatedImage list
This field is required for proper testing on disconnected
environments (for building the list of images to mirror for this
bundle).

The ClusterPolicy images (tags) are also turned into SHA, for the same
reason.

The `operator_bundle_image_builder_script` is transformed from a YAML
ConfigMap into a simple script file, for easier editing.